requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.66.0
anyio==4.10.0
appdirs==1.4.4
arxiv==2.2.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
bidict==0.23.1
bleach==6.2.0
blis==1.3.0
blockbuster==1.5.25
boto3==1.40.31
botocore==1.40.31
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
chainlit==2.7.2
charset-normalizer==3.4.3
ChatterBot==1.2.7
chevron==0.14.0
chromadb==1.0.20
click==8.2.1
cloudpathlib==0.22.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cryptography==44.0.3
cuid==0.4
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
durationpy==0.10
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.116.1
fastjsonschema==2.21.2
feedparser==6.0.12
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
forbiddenfruit==0.1.4
frozenlist==1.7.0
fsspec==2025.9.0
google-auth==2.40.3
google_speech==1.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.29
langchain-core==0.3.75
langchain-experimental==0.3.4
langchain-groq==0.3.7
langchain-openai==0.3.32
langchain-pinecone==0.2.12
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langcodes==3.5.0
langgraph==0.6.6
langgraph-api==0.4.20
langgraph-checkpoint==2.1.1
langgraph-cli==0.4.2
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.12.0
langgraph-sdk==0.2.6
langsmith==0.4.21
language_data==1.3.0
Lazify==0.4.0
literalai==0.1.201
lxml==6.0.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mathparse==0.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.13.1
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j-driver==5.28.2
nest-asyncio==1.6.0
numexpr==2.11.0
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.102.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-alephalpha==0.46.2
opentelemetry-instrumentation-anthropic==0.46.2
opentelemetry-instrumentation-bedrock==0.46.2
opentelemetry-instrumentation-chromadb==0.46.2
opentelemetry-instrumentation-cohere==0.46.2
opentelemetry-instrumentation-crewai==0.46.2
opentelemetry-instrumentation-google-generativeai==0.46.2
opentelemetry-instrumentation-groq==0.46.2
opentelemetry-instrumentation-haystack==0.46.2
opentelemetry-instrumentation-lancedb==0.46.2
opentelemetry-instrumentation-langchain==0.46.2
opentelemetry-instrumentation-llamaindex==0.46.2
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-marqo==0.46.2
opentelemetry-instrumentation-mcp==0.46.2
opentelemetry-instrumentation-milvus==0.46.2
opentelemetry-instrumentation-mistralai==0.46.2
opentelemetry-instrumentation-ollama==0.46.2
opentelemetry-instrumentation-openai==0.46.2
opentelemetry-instrumentation-openai-agents==0.46.2
opentelemetry-instrumentation-pinecone==0.46.2
opentelemetry-instrumentation-qdrant==0.46.2
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-replicate==0.46.2
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-sagemaker==0.46.2
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-instrumentation-threading==0.57b0
opentelemetry-instrumentation-together==0.46.2
opentelemetry-instrumentation-transformers==0.46.2
opentelemetry-instrumentation-urllib3==0.57b0
opentelemetry-instrumentation-vertexai==0.46.2
opentelemetry-instrumentation-watsonx==0.46.2
opentelemetry-instrumentation-weaviate==0.46.2
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.57b0
orjson==3.11.3
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
posthog==3.25.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
psycopg2==2.9.10
PubChemPy==1.0.5
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.4
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
python-vlc==3.0.21203
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
rdkit==2025.3.6
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.1
seaborn==0.13.2
selenium==4.35.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.5.3
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
srsly==2.5.1
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.3
structlog==25.4.0
sympy==1.14.0
syncer==2.0.3
tavily-python==0.7.12
tdc==0.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.0
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traceloop-sdk==0.46.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
truststore==0.10.4
typer==0.17.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
web_cache==1.1.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
